Updated generation threshold to be based on rows generated rather than values
* Also updated command line wording
