+++ generator/src/main/java/com/scottlogic/deg/generator/generation/GenerationConfig.java
@@ -3,14 +3,10 @@ package com.scottlogic.deg.generator.generation;
 import com.scottlogic.deg.generator.generation.combination_strategies.ICombinationStrategy;
 
 public class GenerationConfig {
-    public enum DataGenerationType {
-        FullSequential,
-        Interesting,
-        Random
-    }
 
     private final DataGenerationType dataGenerationType;
     private final ICombinationStrategy combinationStrategy;
+    private final long maxRows = 10_000_000;
 
     public GenerationConfig(
         DataGenerationType dataGenerationType,
@@ -26,5 +22,13 @@ public class GenerationConfig {
 
     public ICombinationStrategy getCombinationStrategy() {
         return combinationStrategy;
+    }
+
+    public long getMaxRows() { return maxRows; }
+
+    public enum DataGenerationType {
+        FullSequential,
+        Interesting,
+        Random
     }
 }
